# backend/src/routes/api/v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Webflow Feedback Tool API",
    "version": "1.0.0",
    "description": "REST API for programmatic access to feedback data, project management, and webhook configuration. Available to Premium and Agency tier subscribers.",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://api.yourapp.com/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "BearerApiKey": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained from POST /api/v1/auth/keys. Use as: Authorization: Bearer wfapi_..."
      },
      "BearerJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token for key management endpoints"
      }
    },
    "schemas": {
...
      "Round": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "frozen", "archived"] },
          "description": { "type": "string", "nullable": true },
          "feedback_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "starts_at": { "type": "string", "format": "date-time", "nullable": true },
          "ends_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/projects/{projectId}/rounds": {
      "get": {
        "summary": "List rounds in project",
        "tags": ["Projects"],
        "parameters": [
          { "name": "projectId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rounds": { "type": "array", "items": { "$ref": "#/components/schemas/Round" } }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}